+++ src/main/java/com/system/libsystem/rest/books/filter/FilterBooksController.java
@@ -21,7 +21,7 @@ public class FilterBooksController {
     }
 
     @GetMapping("/{book_id}")
-    public BookEntity getBookDetails(@PathVariable("book_id") int bookId) {
+    public BookEntity getBookDetails(@PathVariable("book_id") Long bookId) {
         return filterBooksService.getBookDetails(bookId);
     }
